backend_fastapi/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiomysql==0.2.0
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
cos-python-sdk-v5==1.9.27
